vision/faceExpressionClassification_FlyAI/resnet/requirements.txt
Pillow
resnet
torchvision
torch
numpy
flyai
matplotlib
utils
